requirements.txt
faiss-cpu==1.13.2
fastapi==0.135.3
joblib==1.5.3
neurokit2==0.2.13
numpy==2.2.6
openai==2.21.0
pandas==2.3.3
python-dotenv==1.2.2
scipy==1.15.3
uvicorn==0.43.0
xgboost==3.2.0
